Only copy /etc/resolv.conf if instPath != '/'
Fixes a traceback seen where network.py tries to copy /etc/resolv.conf
to /etc/resolv.conf.  The original fix was to get DNS working in %post
scripts during kickstart installs, but that fix assumed that when were
in this block, instPath would be something (such as /mnt/sysimage).

Only copy when instPath is some value other than ''.

+++ network.py
@@ -657,7 +657,7 @@ class Network:
                 self.domains = [domainname]
 
         # /etc/resolv.conf
-        if os.path.isfile('/etc/resolv.conf'):
+        if os.path.isfile('/etc/resolv.conf') and instPath != '':
             destresolv = "%s/etc/resolv.conf" % (instPath,)
             shutil.copy('/etc/resolv.conf', destresolv)
         elif (self.domains != ['localdomain'] and self.domains) or \
